InteractionCoordinates should be an interface
