+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/internal/network/node/importer/FluidHandlerImporterTransferStrategyFactory.java
@@ -6,7 +6,7 @@ import com.refinedmods.refinedstorage2.api.network.node.importer.ImporterTransfe
 import com.refinedmods.refinedstorage2.platform.api.network.node.importer.ImporterTransferStrategyFactory;
 import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.common.internal.storage.channel.StorageChannelTypes;
-import com.refinedmods.refinedstorage2.platform.forge.internal.storage.InteractionCoordinates;
+import com.refinedmods.refinedstorage2.platform.forge.internal.storage.InteractionCoordinatesImpl;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -19,7 +19,7 @@ public class FluidHandlerImporterTransferStrategyFactory implements ImporterTran
                                            final BlockPos pos,
                                            final Direction direction,
                                            final boolean hasStackUpgrade) {
-        final ImporterSource<FluidResource> source = new FluidHandlerImporterSource(new InteractionCoordinates(
+        final ImporterSource<FluidResource> source = new FluidHandlerImporterSource(new InteractionCoordinatesImpl(
             level,
             pos,
             direction
